core: Separate particle storage from particle data
Move particle creation, deletion, storage, cache and tracking logic
to a dedicated source file particle_node.cpp and leave particle
property setters and getters in particle_data.cpp.
