+++ src/core/CMakeLists.txt
@@ -22,6 +22,7 @@ set(EspressoCore_SRC
     npt.cpp
     partCfg_global.cpp
     particle_data.cpp
+    particle_node.cpp
     polymer.cpp
     pressure.cpp
     rattle.cpp
